Move Appbar and its layout wrapper into the app directory
