+++ app/client/components/AppBar.test.js
