+++ app/client/graphql/queries.js
@@ -1,5 +1,6 @@
 import gql from 'graphql-tag'
 
+// TODO should something like this be in the component-model-user package?
 export const CURRENT_USER = gql`
   query CurrentUser {
     currentUser {
