+++ packages/component-elife-ui/client/global/AuthenticatedComponent.js
@@ -3,7 +3,7 @@ import { Redirect } from 'react-router'
 import { Query } from 'react-apollo'
 import Loading from '@elifesciences/component-elife-ui/client/ui/atoms/Loading'
 import { ErrorPage } from '@elifesciences/component-elife-ui/client'
-import { CURRENT_USER } from './queries'
+import { CURRENT_USER } from '../../../../app/client/graphql/queries'
 
 const AuthenticatedComponent = ({ children }) => (
   <Query query={CURRENT_USER}>
