+++ packages/component-elife-ui/client/global/index.js
@@ -1,7 +1,5 @@
-export { default as AppBar } from './AppBar'
 export { default as AuthenticatedComponent } from './AuthenticatedComponent'
 export { default as TwoColumnLayout } from './layout/TwoColumnLayout'
-export { default as Layout } from './layout/Layout'
 export { default as RedirectLayout } from './RedirectLayout'
 export { default as ErrorBoundary } from './ErrorBoundary'
 export { default as CookieNotice } from './CookieNotice'
