+++ packages/component-submission/client/pages/SubmissionWizard/steps/Author/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Query } from 'react-apollo'
-import { CURRENT_USER } from '@elifesciences/component-elife-ui/client/global/queries'
+import { CURRENT_USER } from '../../../../../../../app/client/graphql/queries'
 import AuthorPage from './AuthorPage'
 
 const AuthorPageContainer = ({ handleSubmit, setFieldValue }) => (
